+++ src/BayesFilters/src/UKFCorrection.cpp
@@ -6,6 +6,7 @@
  */
 
 #include <BayesFilters/UKFCorrection.h>
+#include <BayesFilters/utils.h>
 
 using namespace bfl;
 using namespace bfl::sigma_point;
@@ -61,6 +62,21 @@ MeasurementModel& UKFCorrection::getMeasurementModel()
 }
 
 
+std::pair<bool, VectorXd> UKFCorrection::getLikelihood()
+{
+    if ((innovations_.rows() == 0) || (innovations_.cols() == 0))
+        return std::make_pair(false, VectorXd());
+
+    VectorXd likelihood(innovations_.cols());
+    for (std::size_t i = 0; i < innovations_.cols(); i++)
+    {
+        likelihood(i) = utils::multivariate_gaussian_density(innovations_.col(i), VectorXd::Zero(innovations_.rows()), predicted_meas_.covariance(i)).coeff(0);
+    }
+
+    return std::make_pair(true, likelihood);
+}
+
+
 void UKFCorrection::correctStep(const GaussianMixture& pred_state, GaussianMixture& corr_state)
 {
     /* Pick the correct measurement model. */
@@ -80,7 +96,8 @@ void UKFCorrection::correctStep(const GaussianMixture& pred_state, GaussianMixtu
     /* Initialize predicted measurement GaussianMixture. */
     std::pair<std::size_t, std::size_t> meas_sizes = model.getOutputSize();
     std::size_t meas_size = meas_sizes.first + meas_sizes.second;
-    GaussianMixture pred_meas(pred_state.components, meas_size);
+    /* GaussianMixture will effectively resize only if it needs to. */
+    predicted_meas_.resize(pred_state.components, meas_size);
 
     /* Evaluate the joint state-measurement statistics, if possible. */
     bool valid = false;
@@ -94,11 +111,11 @@ void UKFCorrection::correctStep(const GaussianMixture& pred_state, GaussianMixtu
         std::tie(std::ignore, noise_covariance_matrix) = model.getNoiseCovarianceMatrix();
         pred_state_augmented.augmentWithNoise(noise_covariance_matrix);
 
-        std::tie(valid, pred_meas, Pxy) = sigma_point::unscented_transform(pred_state_augmented, ut_weight_, *measurement_model_);
+        std::tie(valid, predicted_meas_, Pxy) = sigma_point::unscented_transform(pred_state_augmented, ut_weight_, *measurement_model_);
     }
     else if (type_ == UKFCorrectionType::Additive)
     {
-        std::tie(valid, pred_meas, Pxy) = sigma_point::unscented_transform(pred_state, ut_weight_, *additive_measurement_model_);
+        std::tie(valid, predicted_meas_, Pxy) = sigma_point::unscented_transform(pred_state, ut_weight_, *additive_measurement_model_);
     }
 
     if (!valid)
@@ -113,8 +130,8 @@ void UKFCorrection::correctStep(const GaussianMixture& pred_state, GaussianMixtu
     /* This temporary is required since some MeasurementModel::innovation methods may try to cast from
        const Ref<const MatrixXd> to MatrixXd resulting in a bfl::any::bad_any_cast.
 
-       Hopefully, using std::move, it is possible to steal the memory from pred_meas.mean(). */
-    MatrixXd y_p = std::move(pred_meas.mean());
+       Hopefully, using std::move, it is possible to steal the memory from predicted_meas_.mean(). */
+    MatrixXd y_p = std::move(predicted_meas_.mean());
     std::tie(valid_innovation, innovation) = model.innovation(y_p, measurement);
 
     if (!valid_innovation)
@@ -124,21 +141,21 @@ void UKFCorrection::correctStep(const GaussianMixture& pred_state, GaussianMixtu
     }
 
     /* Cast innovations once for all. */
-    MatrixXd innovations = any::any_cast<MatrixXd&&>(std::move(innovation));
+    innovations_ = any::any_cast<MatrixXd&&>(std::move(innovation));
 
     /* Process all the components in the mixture. */
     for (size_t i=0; i < pred_state.components; i++)
     {
         /* Evaluate the Kalman Gain
            K = Pxy * (Py)^{-1} */
-        MatrixXd K = Pxy.middleCols(meas_size * i, meas_size) * pred_meas.covariance(i).inverse();
+        MatrixXd K = Pxy.middleCols(meas_size * i, meas_size) * predicted_meas_.covariance(i).inverse();
 
         /* Evaluate the filtered mean.
            x_{k}+ = x{k}- + K * innovation */
-        corr_state.mean(i) = pred_state.mean(i) + K * innovations.col(i);
+        corr_state.mean(i) = pred_state.mean(i) + K * innovations_.col(i);
 
         /* Evaluate the filtered covariance
            P_{k}+ = P_{k}- - K * Py * K' */
-        corr_state.covariance(i) = pred_state.covariance(i) - K * pred_meas.covariance(i) * K.transpose();
+        corr_state.covariance(i) = pred_state.covariance(i) - K * predicted_meas_.covariance(i) * K.transpose();
     }
 }